requirements.txt
pytorch
numpy
cv2
imageio
